This change moves a relatively expensive operation that loads
workflow configuration from the inside of a loop that gets executed
many times and runs it only once afterwards.

Invoking load_workflows reloads workflows for all existing admin_sets.
If we do this after each new admin set, the first call runs for 1 set,
the second call for 2 sets, the third for 3 sets, etc. resulting in
1+2+3+...+n reloads.  This exponential performance gives us a cost of
n(n+1)/2 for the whole process.

Since we're not running any other operations in between, we can just
reload all the workflows after we've created all the admin sets for
a linear cost of n (the number of admin sets) for the whole process.

The intent of this change is to improve performance without changing
behavior, so the test suite should run (faster) without needing to
be modified.

**BEFORE**
Test suite run statistics on Mark's local system
```
Finished in 62 minutes 59 seconds (files took 18.18 seconds to load)
710 examples, 0 failures, 9 pending

Randomized with seed 41307
```

**AFTER**
```
Finished in 30 minutes 34 seconds (files took 17.01 seconds to load)
710 examples, 0 failures, 9 pending

Randomized with seed 4098
```