@mark-dce 2021-03-15
Capistrano Deployment

**BEFORE ~7 minutes**
```
06:56 deploy:log_revision
      01 echo "Branch main (at 7b65ef2d51e9c8f1c556e4c96f6f6b3046905bb7) deployed as release 20210315023628 by mark" >…
    ✔ 01 deploy@qa-etd.library.emory.edu 0.103s
```

**AFTER ~ 4 minutes**
```
03:59 deploy:log_revision
      01 echo "Branch debug-qa (at a2f4ebda6c5e87c4301bd00ac29ebeccfcec9baf) deployed as release 20210315024627 by mar…
    ✔ 01 deploy@qa-etd.library.emory.edu 0.105s
```